refactor(web): adds intermediate DOM-aware OutputTarget abstraction
